# backend/src/swaggerSpec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cultural Sites API",
    "version": "1.0.0",
    "description": "API documentation for Cultural Sites"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "location_lat": { "type": "number" },
                  "location_lon": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Error creating user"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Authenticate user and return token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful"
          },
          "400": {
            "description": "Missing credentials"
          },
          "401": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Error logging in"
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error fetching user"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "location_lat": { "type": "number" },
                  "location_lon": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "No fields to update"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Error updating user"
          }
        }
      }
    },
...
  }
}